refactor(generate_types): reorganize generate types files
