+++ bin/src/generate_supabase_types.dart
