+++ lib/src/src.dart
@@ -1,2 +1,1 @@
-export 'generate_supabase_types.dart';
 export 'supabase/supabase.dart';
